cmake-utils.eclass: use a proper if statement

+++ eclass/cmake-utils.eclass
@@ -428,12 +428,12 @@ enable_cmake-utils_src_prepare() {
 enable_cmake-utils_src_configure() {
 	debug-print-function ${FUNCNAME} "$@"
 
-	[[ "${CMAKE_REMOVE_MODULES}" == "yes" ]] && {
+	if [[ "${CMAKE_REMOVE_MODULES}" == "yes" ]] ; then
 		local name
 		for name in ${CMAKE_REMOVE_MODULES_LIST} ; do
 			find "${S}" -name ${name}.cmake -exec rm -v {} + || die
 		done
-	}
+	fi
 
 	_check_build_dir
